+++ compat/spacedb-oxmysql/fxmanifest.lua
@@ -4,7 +4,7 @@ lua54 'yes'
 
 name 'oxmysql'
 author 'Inkwell'
-version '0.2.2'
+version '0.2.3'
 description 'spacedb oxmysql / mysql-async / ghmattimysql compatibility shim'
 
 dependency 'spacedb'